maklorenchik

+++ src/diff_funcs.cpp
@@ -89,7 +89,7 @@ static Node* CopyNode(Node* node_to_copy) {
     copied_version->value = node_to_copy->value;
 
     if (node_to_copy->type == VAR) {
-        int var_len = strlen(node_to_copy->value.var);
+        size_t var_len = strlen(node_to_copy->value.var);
         copied_version->value.var = (char*) calloc(strlen(node_to_copy->value.var) + 1, sizeof(char));
         memcpy(copied_version->value.var, node_to_copy->value.var, sizeof(char) * var_len);
     }
